+++ database/migrations/2017_09_27_023436_create_diseposes_table.php
@@ -17,7 +17,7 @@ class CreateDiseposesTable extends Migration
             $table->increments('id');
             $table->integer('inventory_id')->unsigned();
             $table->date('date');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('cost', 10, 2);
             $table->decimal('total_cost', 10, 2);
             $table->string('reason');
@@ -31,7 +31,7 @@ class CreateDiseposesTable extends Migration
             $table->increments('id');
             $table->integer('inventory_id')->unsigned();
             $table->date('date');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('cost', 10, 2);
             $table->decimal('total_cost', 10, 2);
             $table->string('reason');
