+++ database/migrations/2017_12_11_235527_create_drygoods_table.php
@@ -49,7 +49,7 @@ class CreateDrygoodsTable extends Migration
 
         Schema::create('drygood_stocks', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('quantity')->unsigned();
+            $table->decimal('quantity', 10, 2)->unsigned();
             $table->decimal('price', 10, 2);
             $table->date('received');
             $table->date('expiration');
